chore: moving dirs around
