+++ .github/workflows/test.yml
@@ -44,3 +44,8 @@ jobs:
 
       - name: Run Tests and Linting
         run: npm run test
+
+      - uses: actions/upload-artifact@v3
+        with:
+          name: test-reports
+          path: test/reports/
