+++ .gitignore
@@ -5,3 +5,5 @@ dist
 .nyc_output
 public
 .DS_Store
+test/reports/*
+!test/reports/.keep
