+++ test/config/cucumber.js
@@ -1,0 +1,8 @@
+module.exports = {
+  default: {
+    parallel: 1,
+    format: ['html:test/reports/cucumber-report.html'],
+    publishQuiet: true,
+    paths: ['test/features'],
+  }
+}
